Refactor: next.js 버전 상향 및 폴더구조 수정
