--- frontend/app/head.tsx
@@ -1,11 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>모두의 여행 - 메인페이지</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta name="description" content="Generated by create next app" />
-      <link rel="icon" href="/favicon.ico" />
-      {/* <link rel="stylesheet" href="./styles/slick.css" /> */}
-    </>
-  );
-}
