+++ frontend/src/components/Main/BoardCard.module.css
