+++ frontend/src/components/Main/CourseCard.module.css
