+++ frontend/src/components/Main/CourseCard.tsx
