+++ frontend/src/components/Main/PlaceCard.module.css
