+++ frontend/src/components/Main/RecruitedLabel.tsx
