+++ frontend/src/interfaces/Board.ts
