+++ frontend/src/interfaces/Course.ts
