+++ frontend/src/interfaces/Place.ts
