+++ frontend/src/styles/palette.ts
