+++ frontend/src/styles/slider.css
